support for very large sequence

+++ src/cpath.h
@@ -26,8 +26,8 @@ static inline void          cpath_set_len(struct cpath* cpath, unsigned row, uns
 static inline void          cpath_set_trans(struct cpath* cpath, unsigned row, unsigned state, unsigned long trans);
 void                        cpath_setup(struct cpath* path, uint32_t len);
 
-static inline uint8_t __bits_len(struct cpath const* cpath, unsigned row, unsigned state);
-static inline size_t  __bits_start(struct cpath const* cpath, unsigned row, unsigned state);
+static inline uint8_t  __bits_len(struct cpath const* cpath, unsigned row, unsigned state);
+static inline uint64_t __bits_start(struct cpath const* cpath, unsigned row, unsigned state);
 
 static inline bool cpath_valid(struct cpath const* cpath, unsigned row, unsigned state)
 {
@@ -37,13 +37,13 @@ static inline bool cpath_valid(struct cpath const* cpath, unsigned row, unsigned
 
 static inline unsigned long cpath_get_len(struct cpath const* cpath, unsigned row, unsigned state)
 {
-    size_t start = __bits_start(cpath, row, state);
+    uint64_t start = __bits_start(cpath, row, state);
     return bitarr_get(cpath->bitarr, start + cpath->bits_trans[state], __bits_len(cpath, row, state));
 }
 
 static inline unsigned long cpath_get_trans(struct cpath const* cpath, unsigned row, unsigned state)
 {
-    size_t start = __bits_start(cpath, row, state);
+    uint64_t start = __bits_start(cpath, row, state);
     return bitarr_get(cpath->bitarr, start, cpath->bits_trans[state]);
 }
 
@@ -53,16 +53,15 @@ static inline void cpath_set_invalid(struct cpath* cpath, unsigned row, unsigned
     cpath_set_len(cpath, row, state, invalids[__bits_len(cpath, row, state)]);
 }
 
-#include <stdio.h>
 static inline void cpath_set_len(struct cpath* cpath, unsigned row, unsigned state, unsigned long len)
 {
-    size_t start = __bits_start(cpath, row, state);
+    uint64_t start = __bits_start(cpath, row, state);
     bitarr_set(cpath->bitarr, len, start + cpath->bits_trans[state], __bits_len(cpath, row, state));
 }
 
 static inline void cpath_set_trans(struct cpath* cpath, unsigned row, unsigned state, unsigned long trans)
 {
-    size_t start = __bits_start(cpath, row, state);
+    uint64_t start = __bits_start(cpath, row, state);
     bitarr_set(cpath->bitarr, trans, start, cpath->bits_trans[state]);
 }
 
@@ -71,9 +70,9 @@ static inline uint8_t __bits_len(struct cpath const* cpath, unsigned row, unsign
     return (uint8_t)(cpath->bits_state[state + 1] - cpath->bits_state[state] - cpath->bits_trans[state]);
 }
 
-static inline size_t __bits_start(struct cpath const* cpath, unsigned row, unsigned state)
+static inline uint64_t __bits_start(struct cpath const* cpath, unsigned row, unsigned state)
 {
-    return row * cpath->bits_state[cpath->nstates] + cpath->bits_state[state];
+    return (uint64_t)row * (uint64_t)cpath->bits_state[cpath->nstates] + (uint64_t)cpath->bits_state[state];
 }
 
 #endif
